[#3]feat: filtering 구현, 컴포넌트 합치기
