+++ src/utils/filterEnum.ts
@@ -1,0 +1,16 @@
+export enum Status {
+  FINISHED = '완료',
+  ONGOING ='진행중',
+  NOT_STARTED = '시작안함'
+}
+
+export enum DateType {
+  GoalDate = 'goalDate',
+  CreatedAt ='createdAt',
+  UpdatedAt = 'updatedAt'
+}
+
+export enum IMPORTANCE {
+  true = 1,
+  false = 0,
+}
